Created methods in the game.rb
